refactor(LogFilter): 필터 로직 내부의 로깅 로직을 책임에 맞게 각 클래스로 분리
