+++ backend/src/main/java/kr/momo/config/filter/LogFilter.java
@@ -1,4 +1,4 @@
-package kr.momo.config;
+package kr.momo.config.filter;
 
 import jakarta.servlet.Filter;
 import jakarta.servlet.FilterChain;
@@ -8,14 +8,18 @@ import jakarta.servlet.ServletResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.UUID;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.MDC;
 
 @Slf4j
+@RequiredArgsConstructor
 public class LogFilter implements Filter {
 
     public static final String TRACE_ID = "traceId";
+
+    private final TraceIdGenerator traceIdGenerator;
+    private final LogGenerator logGenerator;
 
     @Override
     public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain)
@@ -24,26 +28,17 @@ public class LogFilter implements Filter {
         HttpServletRequest httpRequest = (HttpServletRequest) servletRequest;
         HttpServletResponse httpResponse = (HttpServletResponse) servletResponse;
 
-        String traceId = generateShortUuid();
+        String traceId = traceIdGenerator.generateShortUuid();
         MDC.put(TRACE_ID, traceId);
 
-        String httpMethod = httpRequest.getMethod();
-        String requestURI = httpRequest.getRequestURI();
-        String remoteAddr = httpRequest.getRemoteAddr();
         long startTime = System.currentTimeMillis();
-        int status = httpResponse.getStatus();
 
         try {
-            log.info("REQUEST [{}][{} {}][{}]", traceId, httpMethod, requestURI, remoteAddr);
+            logGenerator.logRequest(traceId, httpRequest);
             filterChain.doFilter(servletRequest, servletResponse);
         } finally {
-            long duration = System.currentTimeMillis() - startTime;
-            log.info("RESPONSE [{}][{} {}][{} ms][Status: {}]", traceId, httpMethod, requestURI, duration, status);
+            logGenerator.logResponse(traceId, startTime, httpRequest, httpResponse);
             MDC.clear();
         }
-    }
-
-    private String generateShortUuid() {
-        return UUID.randomUUID().toString().split("-", 2)[0];
     }
 }
