+++ backend/src/main/java/kr/momo/config/filter/TraceIdGenerator.java
@@ -1,0 +1,12 @@
+package kr.momo.config.filter;
+
+import java.util.UUID;
+import org.springframework.stereotype.Component;
+
+@Component
+public class TraceIdGenerator {
+
+    public String generateShortUuid() {
+        return UUID.randomUUID().toString().split("-", 2)[0];
+    }
+}
